requirements/base.txt
# Place your base dependencies here
Django==1.8.2
psycopg2==2.6
djangorestframework==3.1.3
django-model-utils==2.2
django-filter==0.10.0
django-oauth-toolkit==0.8.1
drf-extensions==0.2.7
